requirements.txt
asgiref==3.5.2
Django==4.0.4
django-autoslug==1.9.8
django-environ==0.8.1
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
